+++ test/fixtures/package_bootstrap_bundle/core/bootstrap/config/config.exs
@@ -1,0 +1,3 @@
+import Config
+
+config :fixture_bootstrap, required_value: :configured